Refactor all the things. Remove gulp dependencies in favor of NPM scripts, re structure the file hierarchy to be simpler..
- add ESlint,
- remove gulp
- remove some express-generator boilerplate that we don't need
