+++ .gitignore
@@ -1,6 +1,6 @@
 node_modules/
 bower_components/
-build/
+public/dist/
 .sass-cache/
 .DS_Store
 npm-debug.log
