+++ lib/index.js
@@ -6,4 +6,4 @@ router.get('/', (req, res, next) => {
   res.render('index', { title: "Software Discovery Dashboard" });
 });
 
-module.exports = router;
+export default router;
