+++ lib/search.js
@@ -1,6 +1,6 @@
 import express from 'express';
-import { searchFigshare } from '../repository-clients/figshare-client';
-import { getJsonLd } from '../repository-mappers/figshare-mapper';
+import { searchFigshare } from './repository-clients/figshare-client';
+import { getJsonLd } from './repository-mappers/figshare-mapper';
 const router = express.Router();
 
 const jsonLdAttributes = ['name', 'citation', 'author',
