+++ views/index.jade
@@ -121,5 +121,3 @@ block footer
               li: a(href="#") Community
               li.spacer
               li: a(href="#") Collaborate
-
-
